Fixed credits demo and missing reference in options.
