--- samples/grid/basic/credits/demo.css
@@ -1,11 +1,0 @@
-@import url("https://code.highcharts.com/grid/css/grid.css");
-
-#demo {
-    background-color: var(--highcharts-background-color);
-    color: var(--highcharts-neutral-color-100);
-    padding: 8px;
-}
-
-#container {
-    height: 250px;
-}
