+++ samples/grid/pro/credits/demo.css
@@ -1,0 +1,9 @@
+@import url("https://code.highcharts.com/dashboards/css/datagrid.css");
+
+#demo {
+    padding: 8px;
+}
+
+#container {
+    height: 250px;
+}
