+++ samples/grid/pro/credits/demo.details
