+++ samples/grid/pro/credits/demo.html
@@ -1,4 +1,4 @@
-<script src="https://code.highcharts.com/grid/grid-lite.js"></script>
+<script src="https://code.highcharts.com/dashboards/datagrid.js"></script>
 
 <div id="demo">
     <div>
